Resize window, subsample w/o antialiasing

+++ ImageDisplay_C++(visualstudio)/Image.h
@@ -39,7 +39,7 @@ private:
 	char*	ModifiedData;
 	int		ModWidth;
 	int		ModHeight;
-	float   Scale;
+	double   Scale;
 	int		Aliasing;
 	int		windowOverlay;
